# java/socialapp/src/main/java/com/contoso/socialapp/dto/LikeDto.java
package com.contoso.socialapp.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class LikeDto {
    private String postId;
    private String username;
    private String likedAt;
}