# java/socialapp/src/main/java/com/contoso/socialapp/dto/PostRequest.java
package com.contoso.socialapp.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.Data;

@Data
public class PostRequest {

    @NotBlank
    @Size(min = 1, max = 50)
    private String username;

    @NotBlank
    @Size(min = 1, max = 2000)
    private String content;
}